requirements.txt
Django==1.8.5
django-cors-headers==1.1.0
djangorestframework==3.2.4
requests==2.8.0
